+++ src/components/BookList.js
@@ -1,18 +1,11 @@
-import { useContext } from "react";
 import BookShow from "./BookShow";
-import BooksContext from "../context/books";
+import useBooksContext from "../hooks/useContext";
 
-const BookList = ({ books, delBook, editBook }) => {
+const BookList = () => {
+  const { books } = useBooksContext();
   const renderBooks = () =>
     books.map((book) => {
-      return (
-        <BookShow
-          delBook={delBook}
-          editBook={editBook}
-          key={book.id}
-          book={book}
-        />
-      );
+      return <BookShow key={book.id} book={book} />;
     });
 
   return <div className="book-list"> {renderBooks()}</div>;
