+++ src/hooks/useContext.js
@@ -1,0 +1,6 @@
+import { useContext } from "react";
+import BooksContext from "../context/books";
+
+export default function useBooksContext() {
+  return useContext(BooksContext);
+}
